update component-index.html to better handle root selector. move component tests into cypress folder. update tests.
