+++ cypress.config.js
@@ -9,13 +9,13 @@ module.exports = defineConfig({
     },
   },
   e2e: {
-    setupNodeEvents(on, config) {},
-    specPattern: 'cypress/e2e/**/*.spec.{js,ts,jsx,tsx}',
+    setupNodeEvents() {},
+    specPattern: 'cypress/e2e/**/*.cy.{js,ts,jsx,tsx}',
     excludeSpecPattern: ['**/__snapshots__/*', '**/__image_snapshots__/*'],
   },
   component: {
-    setupNodeEvents(on, config) {},
-    specPattern: 'component/**/*.spec.{js,ts,jsx,tsx}',
+    setupNodeEvents() {},
+    specPattern: 'cypress/component/**/*.cy.{js,ts,jsx,tsx}',
     excludeSpecPattern: ['**/__snapshots__/*', '**/__image_snapshots__/*'],
     devServer: {
       framework: 'create-react-app',
