+++ cypress/component/advanced/products.js
