+++ cypress/e2e/calculator.cy.js
