+++ cypress/support/component-index.html
@@ -7,6 +7,6 @@
     <title>Components App</title>
   </head>
   <body>
-    <div data-cy-root></div>
+    <div id="__cy_root" data-cy-root></div>
   </body>
 </html>
