+++ cypress/support/component.js
@@ -12,6 +12,7 @@
 // You can read more here:
 // https://on.cypress.io/configuration
 // ***********************************************************
+import '../..';
 
 // Import commands.js using ES2015 syntax:
 import './commands';
